# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/enrollments/me": {
      "get": {
        "operationId": "EnrollmentsController_getMyEnrollments",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnrollmentResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List own enrollments",
        "tags": [
          "enrollments"
        ]
      }
    },
    "/api/enrollments/{id}/revoke": {
      "patch": {
        "operationId": "EnrollmentsController_revokeEnrollment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Revoke an enrollment (admin)",
        "tags": [
          "enrollments"
        ]
      }
    },
...
    "/api/courses": {
      "get": {
        "operationId": "CoursesController_listPublished",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Keyword matched against the course title. Max 200 characters.",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Sort direction. Default latest.",
            "schema": {
              "type": "string",
              "enum": [
                "latest",
                "oldest"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number.",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size, default 20, max 100.",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CourseCardDto"
                      }
                    },
                    "page": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "total": {
                      "type": "integer"
                    },
                    "query": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "List published courses",
        "tags": [
          "courses"
        ]
      },
      "post": {
        "operationId": "CoursesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseCardDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a draft course (admin)",
        "tags": [
          "courses"
        ]
      }
    },
    "/api/courses/{slug}": {
      "get": {
        "operationId": "CoursesController_getPublishedBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDetailDto"
                }
              }
            }
          }
        },
        "summary": "Read a published course by slug",
        "tags": [
          "courses"
        ]
      }
    },
    "/api/courses/{id}": {
      "patch": {
        "operationId": "CoursesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseCardDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update a course (admin)",
        "tags": [
          "courses"
        ]
      }
    },
    "/api/courses/{id}/publish": {
      "post": {
        "operationId": "CoursesController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseCardDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Publish a draft course (admin)",
        "tags": [
          "courses"
        ]
      }
    },
    "/api/courses/{id}/unpublish": {
      "post": {
        "operationId": "CoursesController_unpublish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseCardDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Unpublish a course back to draft (admin)",
        "tags": [
          "courses"
        ]
      }
    },
...
    "/api/courses/{slug}/lessons": {
      "get": {
        "operationId": "CourseLessonsController_listPublishedByCourseSlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LessonSummaryDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List published lessons of a course",
        "tags": [
          "courses"
        ]
      }
    },
    "/api/courses/{courseId}/lessons": {
      "post": {
        "operationId": "CourseLessonsController_create",
        "parameters": [
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLessonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonAdminDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a draft lesson under a course (admin)",
        "tags": [
          "courses"
        ]
      }
    },
...
  },
  "info": {
    "title": "Xolvon Backend API",
    "description": "HTTP contract for the Xolvon backend (global prefix /api). Auth: `Authorization: Bearer <accessToken>` (JWT, issued by POST /api/auth/register and POST /api/auth/login). Public routes need no credentials; routes listed with bearer security reject missing or invalid tokens with 401.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT access token. Send it as `Authorization: Bearer <accessToken>`."
      }
    },
    "schemas": {
...
      "EnrollmentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Enrollment id (UUID v4)."
          },
          "courseId": {
            "type": "string",
            "description": "Enrolled course id (UUID v4)."
          },
          "orderId": {
            "type": "string",
            "nullable": true,
            "description": "Originating order id."
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "revoked",
              "expired"
            ]
          },
          "grantedAt": {
            "type": "string",
            "description": "ISO 8601 timestamp."
          },
          "expiresAt": {
            "type": "string",
            "nullable": true,
            "description": "ISO 8601 timestamp."
          },
          "createdAt": {
            "type": "string",
            "description": "ISO 8601 timestamp."
          },
          "courseTitle": {
            "type": "string",
            "nullable": true
          },
          "courseSlug": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "courseId",
          "orderId",
          "status",
          "grantedAt",
          "expiresAt",
          "createdAt"
        ]
      },
...
      "LessonSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Lesson id (UUID v4)."
          },
          "title": {
            "type": "string"
          },
          "orderIndex": {
            "type": "number",
            "description": "Sequence position (order_index)."
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published"
            ]
          }
        },
        "required": [
          "id",
          "title",
          "orderIndex",
          "status"
        ]
      },
      "CourseDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Course id (UUID v4)."
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "description": "Price in IDR."
          },
          "thumbnailUrl": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published"
            ]
          },
          "lessons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LessonSummaryDto"
            }
          }
        },
        "required": [
          "id",
          "title",
          "slug",
          "description",
          "price",
          "thumbnailUrl",
          "status",
          "lessons"
        ]
      },
      "CreateCourseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Course title, max 200 characters.",
            "maxLength": 200
          },
          "slug": {
            "type": "string",
            "description": "URL slug: lowercase letters, digits, hyphens. Max 120 characters.",
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "description": "Description, max 5000 characters, sanitized on write.",
            "maxLength": 5000
          },
          "price": {
            "type": "number",
            "description": "Price in IDR, 0-1000000000.",
            "minimum": 0,
            "maximum": 1000000000
          },
          "thumbnailUrl": {
            "type": "string",
            "description": "Thumbnail URL, max 2048 characters.",
            "maxLength": 2048
          }
        },
        "required": [
          "title",
          "slug",
          "description",
          "price"
        ]
      },
      "CourseCardDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Course id (UUID v4)."
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "description": "Price in IDR."
          },
          "thumbnailUrl": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published"
            ]
          }
        },
        "required": [
          "id",
          "title",
          "slug",
          "description",
          "price",
          "thumbnailUrl",
          "status"
        ]
      },
      "UpdateCourseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Course title, max 200 characters.",
            "maxLength": 200
          },
          "slug": {
            "type": "string",
            "description": "URL slug: lowercase letters, digits, hyphens. Max 120 characters.",
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "description": "Description, max 5000 characters, sanitized on write.",
            "maxLength": 5000
          },
          "price": {
            "type": "number",
            "description": "Price in IDR, 0-1000000000.",
            "minimum": 0,
            "maximum": 1000000000
          },
          "thumbnailUrl": {
            "type": "string",
            "description": "Thumbnail URL, max 2048 characters.",
            "maxLength": 2048
          }
        }
      },
...
      "LessonAdminDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Lesson id (UUID v4)."
          },
          "courseId": {
            "type": "string",
            "description": "Owning course id (UUID v4)."
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "videoObjectKey": {
            "type": "string",
            "nullable": true
          },
          "orderIndex": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published"
            ]
          },
          "createdAt": {
            "type": "string",
            "nullable": true,
            "description": "ISO 8601 timestamp."
          }
        },
        "required": [
          "id",
          "courseId",
          "title",
          "content",
          "videoObjectKey",
          "orderIndex",
          "status",
          "createdAt"
        ]
      },
...
      "CreateLessonDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Lesson title, max 200 characters.",
            "maxLength": 200
          },
          "content": {
            "type": "string",
            "description": "Lesson content, max 20000 characters, sanitized on write.",
            "maxLength": 20000
          },
          "orderIndex": {
            "type": "number",
            "description": "Sequence position, 0-1000000.",
            "minimum": 0,
            "maximum": 1000000
          }
        },
        "required": [
          "title",
          "orderIndex"
        ]
      },
...
    }
  }
}